remove getVectorObject and sepate tests
