+++ lib/cvss.js
@@ -1,5 +1,5 @@
 const definitions = require("./cvss_3_0.json");
-const { roundUp, getVectorObject } = require("./util");
+const { roundUp } = require("./util");
 
 /**
  * Creates a new CVSS object
@@ -13,7 +13,13 @@ function CVSS(vector) {
    * @returns {Number} Calculated  Score
    */
   function getScore() {
-    const vectorObject = getVectorObject(vector);
+    const vectorArray = vector.split("/");
+    const vectorObject = {};
+  
+    for (const entry of vectorArray) {
+      const values = entry.split(":");
+      vectorObject[values[0]] = values[1];
+    }
 
     const scopeChanged = vectorObject.S === "C";
     const ISCBase = calculateISCBase(vectorObject);
@@ -59,7 +65,14 @@ function CVSS(vector) {
    * @returns {Number} Temporal  Score
    */
   function getTemporalScore() {
-    const vectorObject = getVectorObject(vector);
+    const vectorArray = vector.split("/");
+    const vectorObject = {};
+  
+    for (const entry of vectorArray) {
+      const values = entry.split(":");
+      vectorObject[values[0]] = values[1];
+    }
+
     const score = getScore();
     
     const eDef = findMetric("E");
