+++ lib/util.js
@@ -9,21 +9,5 @@ function roundUp(num, precision) {
   return Math.ceil(num * precision) / precision;
 }
 
-/**
- * @param {String} vector The inicial vector CVSS
- *
- * @return The object containing all separate CVSS values 
- */
-function getVectorObject(vector) {
-  const vectorArray = vector.split("/");
-  const vectorObject = {};
 
-  for (const entry of vectorArray) {
-    const values = entry.split(":");
-    vectorObject[values[0]] = values[1];
-  }
-
-  return vectorObject;
-}
-
-module.exports = { roundUp, getVectorObject };
+module.exports = { roundUp };
